requirements.txt
asgiref==3.3.1
autopep8==1.5.4
azure-common==1.1.26
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
Django==3.1.4
django-autoslug==1.9.8
django-bootstrap-pagination==1.7.1
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-js-asset==1.2.2
django-multiselectfield==0.1.12
django-ratelimit==3.0.1
django-storages==1.10.1
easy-thumbnails==2.7.1
idna==2.10
Markdown==3.3.3
Pillow==8.0.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
urllib3==1.26.2
